Move db setup out of TestSupport, make non-static
